+++ internal/storage/sql/migrations/000002_projects.up.sql
@@ -3,9 +3,9 @@ CREATE TABLE IF NOT EXISTS projects (
     id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     name VARCHAR(255) NOT NULL,
     public_key TEXT NOT NULL,
-    created_at BIGINT NOT NULL,
-    updated_at BIGINT NOT NULL,
-    deleted_at BIGINT DEFAULT 0
+    created_at INT NOT NULL,
+    updated_at INT NOT NULL,
+    deleted_at INT DEFAULT 0
 );
 
 CREATE INDEX IF NOT EXISTS idx_projects_name ON projects(name);
